+++ sway-core/src/language/ty/mod.rs
@@ -1,3 +1,5 @@
 mod expression;
+mod variable_mutability;
 
 pub use expression::*;
+pub use variable_mutability::*;
