+++ sway-core/src/semantic_analysis/ast_node/expression/typed_expression.rs
@@ -187,10 +187,10 @@ impl ty::TyExpression {
     }
 
     /// gathers the mutability of the expressions within
-    pub(crate) fn gather_mutability(&self) -> VariableMutability {
+    pub(crate) fn gather_mutability(&self) -> ty::VariableMutability {
         match &self.expression {
             ty::TyExpressionVariant::VariableExpression { mutability, .. } => *mutability,
-            _ => VariableMutability::Immutable,
+            _ => ty::VariableMutability::Immutable,
         }
     }
 
@@ -493,7 +493,7 @@ impl ty::TyExpression {
                     expression: ty::TyExpressionVariant::VariableExpression {
                         name: decl_name,
                         span: name.span(),
-                        mutability: VariableMutability::Immutable,
+                        mutability: ty::VariableMutability::Immutable,
                     },
                     span,
                 }
