+++ sway-core/src/semantic_analysis/ast_node/expression/typed_expression/function_application.rs
@@ -57,7 +57,7 @@ pub(crate) fn instantiate_function_application(
 
             // check for matching mutability
             let param_mutability =
-                convert_to_variable_immutability(param.is_reference, param.is_mutable);
+                ty::convert_to_variable_immutability(param.is_reference, param.is_mutable);
             if exp.gather_mutability().is_immutable() && param_mutability.is_mutable() {
                 errors.push(CompileError::ImmutableArgumentToMutableParameter { span: arg.span() });
             }
